+++ .changes/global-shortcut-feature.md
@@ -1,0 +1,7 @@
+---
+"tauri": patch
+"tauri-runtime": minor
+"tauri-runtime-wry": minor
+---
+
+**Breaking change::* Added the `global-shortcut` Cargo feature.
